requirements.txt
absl-py==0.11.0
appdirs==1.4.4
asgiref==3.3.1
astunparse==1.6.3
attrs==20.2.0
awsebcli==3.19.0
bcrypt==3.2.0
blessed==1.17.10
botocore==1.17.63
cached-property==1.5.2
cachetools==4.1.1
cement==2.8.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
cryptography==3.1.1
cycler==0.10.0
dataclasses==0.6
distlib==0.3.1
Django==3.1.3
docker==4.3.1
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
facenet-pytorch==2.5.1
filelock==3.0.12
future==0.16.0
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
idna==2.7
image==1.5.33
importlib-metadata==2.0.0
jmespath==0.10.0
jsonschema==3.2.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-vggface==0.6
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
mtcnn==0.1.0
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
pandas==1.1.4
paramiko==2.7.2
pathspec==0.5.9
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.5.4
semantic-version==2.5.0
six==1.15.0
sqlparse==0.4.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
texttable==1.6.3
torch==1.7.0
torchvision==0.8.1
typing-extensions==3.7.4.3
urllib3==1.24.3
virtualenv==20.0.34
wcwidth==0.1.9
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.3.0
